계산기Lv.2

+++ src/main/java/org/example/CalculatorLv2.java
@@ -12,13 +12,13 @@ public class CalculatorLv2 {
             System.out.println("계산기를 시작합니다.");
 
             System.out.println("첫번째 수를 입력하세요 : ");
-            int num1 = scanner.nextInt();
+            double num1 = scanner.nextInt();
 
             System.out.println("사칙연산 기호를 입력하세요 : ");
             char operations = scanner.next().charAt(0);
 
             System.out.println("두번째 수를 입력하세요 : ");
-            int num2 = scanner.nextInt();
+            double num2 = scanner.nextInt();
 
             double result = cal.calculate(num1, num2, operations);
             System.out.println("결과는 " + result + "입니다.");
